+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom fields to Job Descriptions
 - Add custom data to job description search results
 - Add international_country_code to phone number details
+- Add customData field to JD search param/result serializers
 
 ## [4.3.0] - 2023-04-27
 ### Changed